+++ material.cpp
@@ -4,7 +4,7 @@ Material::Material( void )
 { }
 
 Material::Material( const Lambertian &bxdf, 
-                    const glm::vec3 &emitted ) :
+                    const glm::dvec3 &emitted ) :
         bxdf_{ bxdf },
         emitted_{ emitted }
 { }
