+++ swpathtracer.cpp
@@ -4,18 +4,17 @@ int main( int num_args, char **arg_vector )
 {
     // Create only one instance of the RNG engine per thread. No more instances
     // are created during the execution of the program.
-    RNG< std::uniform_real_distribution, float, std::mt19937 > rng{ 0.0f, 1.0f };
+    RNG< std::uniform_real_distribution, double, std::mt19937 > rng{ 0.0, 1.0 };
 
     Camera *camera = nullptr;
     Sampler *sampler = nullptr;
     Scene scene{};
     Buffer *rendering_buffer = nullptr;
-    glm::vec3 background_color;
+    glm::dvec3 background_color;
     std::size_t max_path_depth;
     std::string output_filename;
 
     CmdLineParser cmdlineparser( num_args, arg_vector );
-
 
     LuaBind lua( cmdlineparser.getInputScriptFilename() );
 
@@ -104,3 +103,4 @@ int main( int num_args, char **arg_vector )
 
     return EXIT_SUCCESS;
 }
+
